feat(account): 实现交易账户核心功能和模拟交易系统
- 添加账户读写锁机制保障并发安全
- 实现账户余额管理和持仓状态存储
- 新增实盘账户(LiveAccount)和模拟账户(PaperAccount)实现
- 重构交易日志和账户快照功能
- 添加交易监听器和日志记录功能
- 实现交易量计算工具和手续费计算
- 优化持仓策略和交易执行逻辑
- 统一交易接口参数并完善错误处理

+++ clever-task-example/src/main/java/org/clever/quant/Position.java
@@ -17,11 +17,11 @@ public class Position {
     /**
      * 持有总量
      */
-    private long volume;
+    private int volume;
     /**
      * 可用量
      */
-    private long availableVolume;
+    private int availableVolume;
     /**
      * 平均成本价
      */
